+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 * *(pushrules)* Added support for `sender_notification_permission` condition
   kind (used for `@room` mentions).
 * *(crypto)* Added support for `json.RawMessage` in `EncryptMegolmEvent`.
+* *(mediaproxy)* Added `GetMediaResponseCallback` to write proxied response
+  directly instead of having to use an `io.Reader`.
 
 [MSC2781]: https://github.com/matrix-org/matrix-spec-proposals/pull/2781